+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added support for `SourceDescription.poster` for Android & iOS.
+- Made the intervals for the forward and backward skip buttons on the iOS lockScreen configurable.
 
 ## [7.3.0] - 24-06-03